- Added metadata table and refactored task table
