+++ server/src/entities/reminder/index.ts
@@ -1,2 +1,1 @@
 export * from './Reminder.entity'
-export * from './Reminder.options'
