--- server/src/entities/repeatingTaskInstance/RepeatingTaskInstance.options.ts
@@ -1,7 +1,0 @@
-export const IsCheckedOptions: any = {
-    default: false,
-}
-
-export const DateOptions: any = {
-    type: 'timestamptz',
-}
