--- server/src/entities/repeatingTaskInstance/index.ts
@@ -1,1 +1,0 @@
-export * from './RepeatingTaskInstance.entity'
