+++ server/src/entities/task/Task.entity.ts
@@ -1,19 +1,8 @@
-import { BaseEntity, Column, Entity, JoinColumn, ManyToOne, OneToMany, PrimaryGeneratedColumn } from 'typeorm'
+import { BaseEntity, Column, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from 'typeorm'
 
-import { RepeatingTaskInstanceEntity } from '../repeatingTaskInstance'
 import { TaskCardEntity } from '../taskCard'
-import {
-    CheckedOptions,
-    DateOptions,
-    EndDateOptions,
-    IsHabitOptions,
-    IsRepeatingOptions,
-    NextRepeatingInstanceOptions,
-    NoteOptions,
-    RRuleOptions,
-    TaskCardIdOptions,
-    TitleOptions,
-} from './Task.options'
+import { TaskMetaDataEntity } from '../taskMetaData'
+import { CheckedOptions, NoteOptions, TaskCardIdOptions, TitleOptions } from './Task.options'
 
 @Entity('task')
 export class TaskEntity extends BaseEntity {
@@ -26,32 +15,14 @@ export class TaskEntity extends BaseEntity {
     @Column(NoteOptions)
     note: string
 
-    @Column(DateOptions)
-    date: Date
-
-    @Column(EndDateOptions)
-    endDate: Date
-
     @Column(CheckedOptions)
     checked: boolean
 
-    @Column(RRuleOptions)
-    rrule: string
-
-    @Column(IsRepeatingOptions)
-    isRepeating: boolean
-
-    @Column(IsHabitOptions)
-    isHabit: boolean
-
-    @Column(NextRepeatingInstanceOptions)
-    nextRepeatingInstance: Date | null // If task repeating, it holds its next instance
+    @ManyToOne(() => TaskMetaDataEntity, taskMetaData => taskMetaData.tasks)
+    metaData: TaskMetaDataEntity
 
     @Column(TaskCardIdOptions)
     @ManyToOne(() => TaskCardEntity, taskCard => taskCard.tasks, { cascade: true })
     @JoinColumn({ name: 'taskCardId' })
     taskCardId: TaskCardEntity
-
-    @OneToMany(() => RepeatingTaskInstanceEntity, repeatingTaskInstance => repeatingTaskInstance.taskId)
-    repeatingTaskInstances: RepeatingTaskInstanceEntity[]
 }
