+++ server/src/entities/task/Task.options.ts
@@ -13,32 +13,6 @@ export const CheckedOptions: any = {
     default: false,
 }
 
-export const DateOptions: any = {
-    type: 'timestamptz',
-}
-
-export const EndDateOptions: any = {
-    type: 'timestamptz',
-    nullable: true,
-}
-
-export const RRuleOptions: any = {
-    nullable: true,
-}
-
-export const IsRepeatingOptions: any = {
-    default: false,
-}
-
-export const IsHabitOptions: any = {
-    default: false,
-}
-
-export const NextRepeatingInstanceOptions: any = {
-    type: 'timestamptz',
-    nullable: true,
-}
-
 export const TaskCardIdOptions: any = {
     type: 'varchar',
 }
