+++ server/src/entities/task/index.ts
@@ -1,2 +1,1 @@
 export * from './Task.entity'
-export * from './Task.options'
