+++ server/src/entities/taskCard/index.ts
@@ -1,2 +1,1 @@
 export * from './TaskCard.entity'
-export * from './TaskCard.options'
