+++ server/src/entities/taskMetaData/index.ts
@@ -1,0 +1,1 @@
+export * from './TaskMetaData.entity'
