+++ server/src/resolvers/index.ts
@@ -1,7 +1,6 @@
 import { GraphQLDateTime } from 'graphql-iso-date'
 
 import { reminderResolver } from './Reminder/reminder.resolver'
-import { repeatingTaskInstanceResolver } from './RepeatingTaskInstance/RepeatingTaskInstance.resolver'
 import { taskResolver } from './Task/task.resolver'
 import { taskCardResolver } from './TaskCard/taskCard.resolver'
 import { userResolver } from './User/user.resolver'
@@ -11,7 +10,6 @@ export const resolvers = [
     reminderResolver,
     taskResolver,
     taskCardResolver,
-    repeatingTaskInstanceResolver,
 
     // Custom scalars
     { GraphQLDateTime },
